Completed work on "new-style" classes
As documented in the code (see the docstring for the parse() function), these classes hide the XML attributes in class attributes. Additionally, the classes support some "pretty printing" functionality (basically, converting the objects to simple objects that pprint() will handle correctly).

Also, this fixes a bug in the previous work on index_keys and index_keys_compress. If index_keys is set and index_keys_compress is False and two items appeared with the same key, the code would previously raise a ValueError. Now, it creates a list. (We can argue over whether that is a "bug"; however, the behavior was inconsistent with the behavior when index_keys_compress was True.)

Finally, this changes the behavior of _emit() so it will handle the case where index_keys_compress is False and we end up with a list (due to the circumstances just described).