Added an lxml parser
The lxml parser will take an lxml Element or ElementTree and convert it to a dictionary. This is useful to produce consistent dictionaries across modules that return raw XML and lxml Element or ElementTree classes.

Also, fixed a bug that prevented attributes from being emited for CDATA nodes when unparse() was called on new-style elements.