Modifications to the LXMLParser class argument handling
1. Test creating an instance of the _DictSAXHandler class at invocation. This will cause an immediate exception if there is a problem with the arguments.
2. Allow individual arguments to be overridden when the __call__ method is executed.

(There may be further work needed here. The original idea was to allow this to be used as a callback function in various circumstances; hence, the allowance for extra arguments. I'm not sure that ever would have worked correctly. I am especially unsure now that we are passing all the kwargs on to the _DictSAXHandler class.)