Modified the generator test case that checked for "partial processing".
There is a test case that makes sure that the generator will return periodically as it processes an input file. However, this is only really necessary if the Expat processor would call the callbacks as it processes an input file. As long as the two behaviors are consistent, the generator should be as efficient as the callback behavior.

I added a check that tests whether Expat calls its handler routines as it is reading a file, or only once it is done reading the whole file. If it calls the handler routines as it is reading the file, the test ensures that the generator will return as it is reading the file. Otherwise, this test is skipped.